+++ pkg/config/controller_config.go
@@ -16,12 +16,6 @@ const (
 	flagTargetgroupBindingMaxConcurrentReconciles = "targetgroupbinding-max-concurrent-reconciles"
 	defaultLogLevel                               = "info"
 	defaultMaxConcurrentReconciles                = 3
-	// High enough QPS to fit all expected use cases. QPS=0 is not set here, because
-	// client code is overriding it.
-	defaultQPS = 1e6
-	// High enough Burst to fit all expected use cases. Burst=0 is not set here, because
-	// client code is overriding it.
-	defaultBurst = 1e6
 )
 
 // ControllerConfig contains the controller configuration
@@ -62,45 +56,5 @@ func (cfg *ControllerConfig) Validate() error {
 	if len(cfg.ClusterName) == 0 {
 		return fmt.Errorf("Kubernetes cluster name must be specified")
 	}
-	if cfg.RuntimeConfig.Scheme == nil {
-		return fmt.Errorf("Controller runtime scheme not initialzied")
-	}
 	return nil
-}
-
-func buildRestConfig(masterURL, kubeconfigPath string) (*rest.Config, error) {
-	if kubeconfigPath == "" && masterURL == "" {
-		kubeconfig, err := rest.InClusterConfig()
-		if err == nil {
-			return kubeconfig, nil
-		}
-	}
-	return clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
-		&clientcmd.ClientConfigLoadingRules{ExplicitPath: kubeconfigPath},
-		&clientcmd.ConfigOverrides{ClusterInfo: clientcmdapi.Cluster{Server: masterURL}}).ClientConfig()
-}
-
-// BuildRestConfig builds the REST config for the controller runtime
-func BuildRestConfig(rtCfg RuntimeConfig) (*rest.Config, error) {
-	restCfg, err := buildRestConfig(rtCfg.APIServer, rtCfg.KubeConfig)
-	if err != nil {
-		return nil, err
-	}
-	restCfg.QPS = defaultQPS
-	restCfg.Burst = defaultBurst
-	return restCfg, nil
-}
-
-// BuildRuntimeOptions builds the options for the controller runtime based on config
-func BuildRuntimeOptions(rtCfg RuntimeConfig) ctrl.Options {
-	return ctrl.Options{
-		Scheme:                  rtCfg.Scheme,
-		Port:                    rtCfg.ControllerPort,
-		MetricsBindAddress:      rtCfg.MetricsBindAddress,
-		LeaderElection:          rtCfg.EnableLeaderElection,
-		LeaderElectionID:        rtCfg.LeaderElectionID,
-		LeaderElectionNamespace: rtCfg.LeaderElectionNamespace,
-		Namespace:               rtCfg.WatchNamespace,
-		SyncPeriod:              &rtCfg.SyncPeriod,
-	}
 }
